Create README.md
Python_Practice_Files
=====================
These are programs I wrote while learning Python. Python is an easy-to-use language that can be powerful. Unlike C++, it is NOT a compiled language, but is interpreted at run-time.

While no ground-breaking innovations are likely to be found in this repository, they can serve as examples of programs I wrote or modified while learning this new programming resource! If anyone wants to look at examples of my python work, check out these files...

Thank you!
Dan